test: split codex facade coverage by module
