+++ src/codex-prompt.test.ts
@@ -172,3 +172,445 @@ test("buildCodexResumePrompt falls back to current failure context when the jour
   assert.match(prompt, /Command\/source: npm run build/);
   assert.match(prompt, /Respond in this exact footer format at the end:/);
 });
+
+test("buildCodexPrompt emphasizes compressed local-review root causes during local_review_fix", () => {
+  const failureContext: FailureContext = {
+    category: "blocked",
+    summary: "Local review found 2 actionable findings across 2 root causes.",
+    signature: "local-review:high:high:2:1:clean",
+    command: null,
+    details: ["findings=2", "root_causes=2", "summary=/tmp/reviews/issue-46/head-deadbeef.md"],
+    url: null,
+    updated_at: "2026-03-12T00:00:00Z",
+  };
+
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "local_review_fix" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    failureContext,
+    localReviewRepairContext: {
+      summaryPath: "/tmp/reviews/issue-46/head-deadbeef.md",
+      findingsPath: "/tmp/reviews/issue-46/head-deadbeef.json",
+      relevantFiles: ["src/supervisor.ts", "src/codex.ts"],
+      rootCauses: [
+        {
+          severity: "high",
+          summary: "State inference sends local-review retries through implementing instead of a dedicated fix mode.",
+          file: "src/supervisor.ts",
+          lines: "761-771",
+        },
+        {
+          severity: "medium",
+          summary: "Prompt guidance still frames the turn as checkpoint maintenance.",
+          file: "src/codex.ts",
+          lines: "92-101",
+        },
+      ],
+      priorMissPatterns: [
+        {
+          fingerprint: "src/supervisor.ts|retry-mode",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/supervisor.ts",
+          line: 761,
+          summary: "Repair retries can loop through the wrong state.",
+          rationale: "A prior external review caught a retry path that stayed in implementing instead of a dedicated repair state.",
+          sourceArtifactPath: "/tmp/reviews/issue-46/external-review-misses-head-old.json",
+          sourceHeadSha: "oldhead123",
+          lastSeenAt: "2026-03-12T00:00:00Z",
+        },
+      ],
+      verifierGuardrails: [
+        {
+          id: "retry-mode",
+          title: "Re-check retry mode state handoff",
+          file: "src/supervisor.ts",
+          line: 761,
+          summary: "Confirm the repair retry path enters the dedicated local_review_fix state instead of implementing.",
+          rationale: "A prior verifier miss cleared a retry transition without checking the state update branch directly.",
+        },
+      ],
+    },
+  });
+
+  assert.match(prompt, /Supervisor state: local_review_fix/);
+  assert.match(prompt, /blocking the PR or merge/);
+  assert.match(prompt, /Relevant files to inspect first:/);
+  assert.match(prompt, /src\/supervisor\.ts/);
+  assert.match(prompt, /State inference sends local-review retries/);
+  assert.match(prompt, /Committed regression-oriented guardrails:/);
+  assert.match(prompt, /Repair retries can loop through the wrong state\./);
+  assert.match(prompt, /Committed verifier guardrails:/);
+  assert.match(prompt, /Re-check retry mode state handoff/);
+});
+
+test("buildCodexPrompt suppresses flat next-action bullet lists during local_review_fix", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "local_review_fix" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    journalExcerpt: `# Issue #46: Add a dedicated local_review_fix repair mode
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: The repair path still carries forward stale checkpoint guidance.
+- Next 1-3 actions:
+- Leave the PR as a stable checkpoint for handoff.
+* Re-read the broad implementation plan before touching code.
+
+### Scratchpad
+- Keep this section short.`,
+    localReviewRepairContext: {
+      summaryPath: "/tmp/reviews/issue-46/head-deadbeef.md",
+      findingsPath: "/tmp/reviews/issue-46/head-deadbeef.json",
+      relevantFiles: ["src/codex.ts"],
+      rootCauses: [
+        {
+          severity: "high",
+          summary: "Prompt guidance should ignore stale checkpoint-maintenance handoff text during repair.",
+          file: "src/codex.ts",
+          lines: "1-200",
+        },
+      ],
+      priorMissPatterns: [],
+      verifierGuardrails: [],
+    },
+  });
+
+  assert.doesNotMatch(prompt, /Leave the PR as a stable checkpoint for handoff\./);
+  assert.doesNotMatch(prompt, /Re-read the broad implementation plan before touching code\./);
+  assert.match(prompt, /### Current Handoff/);
+  assert.match(prompt, /### Scratchpad/);
+});
+
+test("buildCodexPrompt keeps explicit operator overrides during local_review_fix", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "local_review_fix" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    journalExcerpt: `# Issue #46: Add a dedicated local_review_fix repair mode
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: The repair path still carries forward stale checkpoint guidance.
+- Operator override: Keep the temporary compatibility shim in place while fixing the review finding.
+- Next 1-3 actions:
+  - Leave the PR as a stable checkpoint for handoff.
+
+### Scratchpad
+- Keep this section short.`,
+    localReviewRepairContext: {
+      summaryPath: "/tmp/reviews/issue-46/head-deadbeef.md",
+      findingsPath: "/tmp/reviews/issue-46/head-deadbeef.json",
+      relevantFiles: ["src/codex.ts"],
+      rootCauses: [
+        {
+          severity: "high",
+          summary: "Prompt guidance should ignore stale checkpoint-maintenance handoff text during repair.",
+          file: "src/codex.ts",
+          lines: "1-200",
+        },
+      ],
+      priorMissPatterns: [],
+      verifierGuardrails: [],
+    },
+  });
+
+  assert.match(prompt, /Operator override: Keep the temporary compatibility shim in place/);
+  assert.doesNotMatch(prompt, /Leave the PR as a stable checkpoint for handoff\./);
+});
+
+test("buildCodexPrompt suppresses stale handoff next actions during addressing_review", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "addressing_review" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    journalExcerpt: `# Issue #46: Add a dedicated review-repair mode
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: Review-thread fixes should override stale checkpoint advice.
+- Next 1-3 actions:
+  - Leave the PR as a stable checkpoint for handoff.
+  - Re-read the broad implementation plan before touching code.
+
+### Scratchpad
+- Keep this section short.`,
+    externalReviewMissContext: {
+      artifactPath: "/tmp/reviews/issue-46/external-review-misses-head-deadbeef.json",
+      matchedCount: 0,
+      nearMatchCount: 1,
+      missedCount: 1,
+      regressionTestCandidates: [],
+      missedFindings: [
+        {
+          sourceKind: "review_thread",
+          sourceId: "thread-1",
+          sourceUrl: "https://example.test/thread-1#comment-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/codex.ts",
+          line: 180,
+          summary: "Live review guidance should take priority over stale handoff steps.",
+          rationale: "The active thread already describes the blocker to resolve before merge.",
+          url: "https://example.test/thread-1#comment-1",
+        },
+      ],
+    },
+  });
+
+  assert.doesNotMatch(prompt, /Leave the PR as a stable checkpoint for handoff\./);
+  assert.doesNotMatch(prompt, /Re-read the broad implementation plan before touching code\./);
+  assert.match(prompt, /Review threads are the primary task\./);
+  assert.match(prompt, /Live review guidance should take priority over stale handoff steps\./);
+});
+
+test("buildCodexPrompt keeps explicit operator overrides during addressing_review", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "addressing_review" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    journalExcerpt: `# Issue #46: Add a dedicated review-repair mode
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: Review-thread fixes should override stale checkpoint advice.
+- Operator override: Keep the temporary compatibility shim in place while resolving the review comment.
+- Next 1-3 actions:
+  - Leave the PR as a stable checkpoint for handoff.
+
+### Scratchpad
+- Keep this section short.`,
+    externalReviewMissContext: {
+      artifactPath: "/tmp/reviews/issue-46/external-review-misses-head-deadbeef.json",
+      matchedCount: 0,
+      nearMatchCount: 0,
+      missedCount: 1,
+      regressionTestCandidates: [],
+      missedFindings: [
+        {
+          sourceKind: "review_thread",
+          sourceId: "thread-1",
+          sourceUrl: "https://example.test/thread-1#comment-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/codex.ts",
+          line: 180,
+          summary: "Live review guidance should take priority over stale handoff steps.",
+          rationale: "The active thread already describes the blocker to resolve before merge.",
+          url: "https://example.test/thread-1#comment-1",
+        },
+      ],
+    },
+  });
+
+  assert.match(prompt, /Operator override: Keep the temporary compatibility shim in place/);
+  assert.doesNotMatch(prompt, /Leave the PR as a stable checkpoint for handoff\./);
+});
+
+test("buildCodexPrompt suppresses stale handoff next actions during repairing_ci", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "repairing_ci" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    journalExcerpt: `# Issue #46: Add a dedicated repair mode
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: CI repair should override stale checkpoint advice.
+- Next exact step: Leave the PR as a stable checkpoint for handoff.
+- Primary failure or risk: npm run build currently fails on the active branch.
+
+### Scratchpad
+- Keep this section short.`,
+  });
+
+  assert.doesNotMatch(prompt, /Leave the PR as a stable checkpoint for handoff\./);
+  assert.match(prompt, /- Next exact step: suppressed during active CI repair/);
+  assert.match(prompt, /- Primary failure or risk: npm run build currently fails on the active branch\./);
+});
+
+test("buildCodexPrompt suppresses structured next exact step guidance during local_review_fix without dropping later fields", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "local_review_fix" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    journalExcerpt: `# Issue #46: Add a dedicated local_review_fix repair mode
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: The repair path still carries forward stale checkpoint guidance.
+- Next exact step: Leave the PR as a stable checkpoint for handoff.
+- Verification gap: Full npm test was not rerun yet.
+
+### Scratchpad
+- Keep this section short.`,
+    localReviewRepairContext: {
+      summaryPath: "/tmp/reviews/issue-46/head-deadbeef.md",
+      findingsPath: "/tmp/reviews/issue-46/head-deadbeef.json",
+      relevantFiles: ["src/codex.ts"],
+      rootCauses: [
+        {
+          severity: "high",
+          summary: "Prompt guidance should ignore stale checkpoint-maintenance handoff text during repair.",
+          file: "src/codex.ts",
+          lines: "1-200",
+        },
+      ],
+      priorMissPatterns: [],
+      verifierGuardrails: [],
+    },
+  });
+
+  assert.doesNotMatch(prompt, /Leave the PR as a stable checkpoint for handoff\./);
+  assert.match(prompt, /- Next exact step: suppressed during active local-review repair/);
+  assert.match(prompt, /- Verification gap: Full npm test was not rerun yet\./);
+});
+
+test("buildCodexPrompt surfaces saved external review misses during addressing_review", () => {
+  const prompt = buildCodexPrompt({
+    repoSlug: "owner/repo",
+    issue,
+    branch: "codex/issue-46",
+    workspacePath: "/tmp/workspaces/issue-46",
+    state: "addressing_review" satisfies RunState,
+    pr: null,
+    checks: [],
+    reviewThreads: [],
+    alwaysReadFiles: [],
+    onDemandMemoryFiles: [],
+    journalPath: "/tmp/workspaces/issue-46/.codex-supervisor/issue-journal.md",
+    externalReviewMissContext: {
+      artifactPath: "/tmp/reviews/issue-46/external-review-misses-head-deadbeef.json",
+      matchedCount: 0,
+      nearMatchCount: 1,
+      missedCount: 4,
+      regressionTestCandidates: [
+        {
+          id: "src/auth.ts|42|permission-guard-is-bypassed",
+          title: "Add regression coverage for Permission guard is bypassed",
+          file: "src/auth.ts",
+          line: 42,
+          summary: "Permission guard is bypassed.",
+          rationale: "This fallback skips the permission guard and lets unauthorized callers update records.",
+          sourceKind: "review_thread",
+          sourceId: "thread-1",
+          sourceThreadId: "thread-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          sourceUrl: "https://example.test/thread-1#comment-1",
+          qualificationReasons: ["missed_by_local_review", "non_low_severity", "high_confidence", "file_scoped", "line_scoped"],
+        },
+      ],
+      missedFindings: [
+        {
+          sourceKind: "review_thread",
+          sourceId: "thread-1",
+          sourceUrl: "https://example.test/thread-1#comment-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/auth.ts",
+          line: 42,
+          summary: "Permission guard is bypassed.",
+          rationale: "This fallback skips the permission guard and lets unauthorized callers update records.".repeat(20),
+          url: "https://example.test/thread-1#comment-1",
+        },
+        {
+          sourceKind: "review_thread",
+          sourceId: "thread-2",
+          sourceUrl: "https://example.test/thread-2#comment-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/auth.ts",
+          line: 43,
+          summary: "Second finding",
+          rationale: "Second rationale",
+          url: "https://example.test/thread-2#comment-1",
+        },
+        {
+          sourceKind: "review_thread",
+          sourceId: "thread-3",
+          sourceUrl: "https://example.test/thread-3#comment-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/auth.ts",
+          line: 44,
+          summary: "Third finding",
+          rationale: "Third rationale",
+          url: "https://example.test/thread-3#comment-1",
+        },
+        {
+          sourceKind: "review_thread",
+          sourceId: "thread-4",
+          sourceUrl: "https://example.test/thread-4#comment-1",
+          reviewerLogin: "copilot-pull-request-reviewer",
+          file: "src/auth.ts",
+          line: 45,
+          summary: "Fourth finding",
+          rationale: "Fourth rationale",
+          url: "https://example.test/thread-4#comment-1",
+        },
+      ],
+    },
+  });
+
+  assert.match(prompt, /External review miss context:/);
+  assert.match(prompt, /matched=0 near_match=1 missed=4/);
+  assert.match(prompt, /Regression-test candidates from confirmed misses:/);
+  assert.match(prompt, /Add regression coverage for Permission guard is bypassed/);
+  assert.match(prompt, /qualified_by=missed_by_local_review, non_low_severity, high_confidence, file_scoped, line_scoped/);
+  assert.match(prompt, /Permission guard is bypassed\./);
+  assert.match(prompt, /copilot-pull-request-reviewer/);
+  assert.match(prompt, /Additional missed findings omitted: 1/);
+  assert.ok(prompt.split("This fallback skips the permission guard").length - 1 <= 3);
+});
